# feat(metainfo): improve performance and reduce memory allocation