--- a/.gitignore
+++ b/.gitignore
@@ -17,3 +17,4 @@
 # IDE or Editor's config
 .idea/
 
+testdata