--- a/cloud/metainfo/utils.go
+++ b/cloud/metainfo/utils.go
@@ -21,7 +21,7 @@ import (
 
 // HasMetaInfo detects whether the given context contains metainfo.
 func HasMetaInfo(ctx context.Context) bool {
-	return getKV(ctx) != nil
+	return getNode(ctx) != nil
 }
 
 // SetMetaInfoFromMap retrieves metainfo key-value pairs from the given map and sets then into the context.
@@ -30,12 +30,34 @@ func SetMetaInfoFromMap(ctx context.Context, m map[string]string) context.Contex
 	if ctx == nil {
 		return nil
 	}
+
+	var n node
 	for k, v := range m {
-		if t, nk := determineKeyType(k, v); t != invalidType {
-			ctx = addKV(ctx, t, nk, v)
+		if len(k) == 0 || len(v) == 0 {
+			continue
+		}
+
+		switch {
+		case strings.HasPrefix(k, PrefixTransientUpstream):
+			if len(k) > lenPTU { // do not move this condition to the case statement to prevent a PTU matches PT
+				n.stale = append(n.stale, kv{key: k[lenPTU:], val: v})
+			}
+		case strings.HasPrefix(k, PrefixTransient):
+			if len(k) > lenPT {
+				n.transient = append(n.transient, kv{key: k[lenPT:], val: v})
+			}
+		case strings.HasPrefix(k, PrefixPersistent):
+			if len(k) > lenPP {
+				n.persistent = append(n.persistent, kv{key: k[lenPP:], val: v})
+			}
 		}
 	}
-	return ctx
+
+	if n.size() == 0 {
+		// TODO: remove this?
+		return ctx
+	}
+	return withNode(ctx, &n)
 }
 
 // SaveMetaInfoToMap set key-value pairs from ctx to m while filtering out transient-upstream data.
@@ -50,33 +72,4 @@ func SaveMetaInfoToMap(ctx context.Context, m map[string]string) {
 	for k, v := range GetAllPersistentValues(ctx) {
 		m[PrefixPersistent+k] = v
 	}
-}
-
-const (
-	lenPTU = len(PrefixTransientUpstream)
-	lenPT  = len(PrefixTransient)
-	lenPP  = len(PrefixPersistent)
-)
-
-// determineKeyType tests whether the given key-value pair is a valid metainfo and returns its info type with a new appropriate key.
-func determineKeyType(k, v string) (infoType infoType, newKey string) {
-	if len(k) == 0 || len(v) == 0 {
-		return invalidType, k
-	}
-
-	switch {
-	case strings.HasPrefix(k, PrefixTransientUpstream):
-		if len(k) > lenPTU {
-			return transientUpstreamType, k[lenPTU:]
-		}
-	case strings.HasPrefix(k, PrefixTransient):
-		if len(k) > lenPT {
-			return transientType, k[lenPT:]
-		}
-	case strings.HasPrefix(k, PrefixPersistent):
-		if len(k) > lenPP {
-			return persistentType, k[lenPP:]
-		}
-	}
-	return invalidType, k
 }